requirement.txt
# conda create -n DiffusionModelFamily python=3.10

# torch
torch==2.3.1
torchvision==0.18.1

# torch util
einops==0.8.0

# diffusers
huggingface-hub==0.25.1
accelerate==1.2.1
diffusers==0.32.1
transformers==4.47.1
tokenizers==0.21.0
sentencepiece==0.2.0
protobuf==3.20.3
safetensors==0.4.5
datasets==3.2.0

# metric
opencv-python==4.11.0.86
hpsv2==1.2.0
pytorch_fid==0.3.0
pytorch_image_generation_metrics==0.6.1
scikit-image==0.25.0
lpips==0.1.4

# Hydra
hydra-core==1.3.2
omegaconf==2.3.0
pickle5==0.0.11
PyYAML==6.0.2

# misc
numpy==2.1.3
matplotlib==3.9.2
pandas==2.2.3
Pillow==10.1.0
tqdm==4.67.1
requests==2.32.3
